requirements.txt
google-cloud-bigquery>=3.19.0
google-cloud-storage>=2.18.2
grpcio-tools>=1.67.1
requests~=2.32.2
v4-proto>=7.0.2
datadog>=0.49.1